v5.6.0 merge part 2

v5.6.0 merge part 3

+++ insts/std/MMOVE/msetup.cpp
@@ -109,8 +109,8 @@ get_rvb_setup_params(double Dimensions[],       /* array of 5 elements */
 int
 get_setup_params(double Dimensions[],       /* array of 5 elements */
 				 AttenuationParams *params,
-				 float *rvb_time,
-                 float  *abs_factor,
+				 double *rvb_time,
+                 double  *abs_factor,
                  int    *UseMikes,
                  double *MikeAngle,
                  double *MikePatternFactor)
